refactored controllers into their own folders
