+++ src/cli/ControllerFactory.cpp
@@ -1,11 +1,11 @@
 #include "ControllerFactory.h"
 #include "IController.h"
-#include "CalibrationController.h"
-#include "ProcessingController.h"
-#include "ScanController.h"
-#include "FilterTestingController.h"
-#include "MoveController.h"
-#include "HomeController.h"
+#include "calibration/CalibrationController.h"
+#include "processing/ProcessingController.h"
+#include "scan/ScanController.h"
+#include "filter_testing/FilterTestingController.h"
+#include "move/MoveController.h"
+#include "home/HomeController.h"
 
 #include "CalibrationFileHandler.h"
 #include "ScanFileHandler.h"
