+++ src/controller/filter_testing/CMakeLists.txt
@@ -1,0 +1,6 @@
+target_sources(swag_scanner_lib PRIVATE
+        ${CMAKE_CURRENT_SOURCE_DIR}/FilterTestingController.cpp
+        ${CMAKE_CURRENT_SOURCE_DIR}/FilterTestingController.h
+        )
+
+target_include_directories(swag_scanner_lib PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
