+++ src/controller/home/CMakeLists.txt
@@ -1,0 +1,6 @@
+target_sources(swag_scanner_lib PRIVATE
+        ${CMAKE_CURRENT_SOURCE_DIR}/HomeController.cpp
+        ${CMAKE_CURRENT_SOURCE_DIR}/HomeController.h
+        )
+
+target_include_directories(swag_scanner_lib PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
